requirements.txt
Django>=5.0
djangorestframework>=3.15
python-dotenv>=1.0
psycopg2-binary>=2.9
gunicorn>=22.0
whitenoise>=6.7
